make two separate tabs to choose between React and HTML on code preview
