+++ apps/web/src/hooks/use-stored-state.ts
@@ -1,0 +1,25 @@
+import * as React from "react";
+
+export const useStoredState = <T extends string | undefined>(
+  key: string,
+  defaultValue: T,
+): [state: T, setState: (newValue: T) => void] => {
+  const [state, setState] = React.useState<T>(defaultValue);
+  React.useEffect(() => {
+    const storedValue = localStorage.getItem(key);
+    if (storedValue) {
+      setState(storedValue as T);
+    }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  return [
+    state,
+    (newValue) => {
+      if (newValue) {
+        localStorage.setItem(key, newValue);
+      }
+      setState(newValue);
+    },
+  ];
+};
